requirements.txt
pandas==1.5.3
numpy==1.20.3
streamlit==1.18.0
matplotlib==3.7.2
urllib3==1.26.9
Pillow==8.4.0
requests==2.27.1
folium==0.12.1
flexpolyline==0.1.0
streamlit-folium==0.4.0
qrcode